+++ src/DoExport.ts
@@ -1,6 +1,7 @@
 import { workspace, window, commands, TextEditor, TextDocument, extensions, OutputChannel, Extension, TextEditorVisibleRangesChangeEvent, WorkspaceConfiguration } from "vscode";
 import { isNullOrUndefined } from "util";
 import path = require("path");
+import { Configuration } from "./util/Configuration";
 
 export class DoExport {
 
@@ -43,7 +44,7 @@ export class DoExport {
     }
 
     private getConfiguration():WorkspaceConfiguration {
-        return workspace.getConfiguration("bdev-al-xml-doc");
+        return workspace.getConfiguration(Configuration.ExtensionIdent());
     }
 
     private VerifyPrerequisite(): boolean {
@@ -104,6 +105,10 @@ export class DoExport {
                     return;
                 }
                 markdownPath = path.join(workspaceRoot, "doc");			
+            } else {
+                if (workspace.workspaceFolders) {
+                    markdownPath = markdownPath.replace('${workspaceFolder}', this.getDirectoryName(workspace.workspaceFolders[0].uri.toString()));
+                }
             }
             console.debug("Using export path: " + markdownPath);
         } catch (ex) {