+++ src/util/ALSyntaxUtil.ts
@@ -1,5 +1,6 @@
 import { isNullOrUndefined } from "util";
 import { TextDocument, Range, Position } from "vscode";
+import { Configuration } from "./Configuration";
 
 export class ALSyntaxUtil {
     public static IsObject(line: string): boolean {
@@ -22,28 +23,54 @@ export class ALSyntaxUtil {
         return true;
     }
 
-    public static IsProcedure(line: string): boolean {
+    public static IsProcedure(line: string, prevLine: string): boolean {
         if (line === null) {
             return false;
         }
 
+        let procedureTypes = Configuration.ProcedureTypes();
+
+        const isEventPublisher: boolean = (prevLine.trim().startsWith('[BusinessEvent')) || (prevLine.trim().startsWith('[IntegrationEvent')) || (prevLine.trim().startsWith('[InternalEvent')) ;
+        if (isEventPublisher) {
+            if ((procedureTypes.length === 0) || (procedureTypes.includes('Event Publisher'))) {
+                return true;
+            }
+            return false;
+        }
+
+        const isEventSubscriber: boolean = prevLine.trim().startsWith('[EventSubscriber');
+        if (isEventSubscriber) {
+            if ((procedureTypes.length === 0) || (procedureTypes.includes('Event Subscriber'))) {
+                return true;
+            }
+            return false;
+        }
+        
+        const isTestProcedure: boolean = prevLine.trim().startsWith('[Test');
+        if (isTestProcedure) {
+            if ((procedureTypes.length === 0) || (procedureTypes.includes('Test Procedures'))) {
+                return true;
+            }
+            return false;
+        }
+
         const isProcedure: boolean = line.trim().startsWith('procedure');
-        if (isProcedure) {
+        if ((isProcedure) && ((procedureTypes.length === 0) || (procedureTypes.includes('Global Procedures')))) {
             return true;
         }
 
         const isInternalProcedure: boolean = line.trim().startsWith('internal procedure');
-        if (isInternalProcedure) {
+        if ((isInternalProcedure) && ((procedureTypes.length === 0) || (procedureTypes.includes('Internal Procedures')))) {
             return true;
         }
 
         const isLocalProcedure: boolean = line.trim().startsWith('local procedure');
-        if (isLocalProcedure) {
+        if ((isLocalProcedure) && ((procedureTypes.length === 0) || (procedureTypes.includes('Local Procedures')))) {
             return true;
         }
 
         const isTriggerProcedure: boolean = line.trim().startsWith('trigger');
-        if (isTriggerProcedure) {
+        if ((isTriggerProcedure) && ((procedureTypes.length === 0) || (procedureTypes.includes('Trigger Procedures')))) {
             return true;
         }
 
@@ -51,7 +78,7 @@ export class ALSyntaxUtil {
     }
 
     public static IsBeginEnd(line: string): boolean {
-        return (line.toLowerCase().match(/\bbegin|\bend/) !== null);
+        return (line.toLowerCase().match(/\bbegin\b|\bend\b/) !== null);
     }
 
     private static AnalyzeDefinition(regExResult: RegExpMatchArray | null) : RegExpMatchArray | null {
@@ -62,7 +89,7 @@ export class ALSyntaxUtil {
     }
 
     public static GetALProcedureState(document: TextDocument, procedureLineNo: number = 0): { name: string; position: Range; definition: { [key: string]: string; }; documentation: string } | null {        
-        let alCode = document.getText().split('\r\n');
+        let alCode = document.getText().replace('\r','').split('\n');
         if (procedureLineNo === 0) {
             procedureLineNo = alCode.length - 1;
         }
@@ -71,7 +98,7 @@ export class ALSyntaxUtil {
             let line = alCode[lineNo];
             switch (true)
             {
-                case ALSyntaxUtil.IsProcedure(line):
+                case ALSyntaxUtil.IsProcedure(line, (lineNo > 0) ? alCode[lineNo - 1] : ""):
                     if (alProcedureState !== null ){
                         return alProcedureState;
                     }
@@ -106,7 +133,7 @@ export class ALSyntaxUtil {
 
     public static FindProcedures(code: string): any {
         let procedures: { procedureName : string, lineNo: number }[] = [];
-        code.match(/(?<!\/\/\/\s*)(((?!local)procedure|(?!internal)procedure)\s+([A-Za-z0-9_]+)\b[^\(]*\)*.+)/g)?.forEach(match => {
+        code.match(/(?<!\/\/\/.*)((procedure)\s+(\")?(.+)\b(\")?[\(]*\)*.+)/g)?.forEach(match => {
             procedures.push({
                 procedureName: match,
                 lineNo: this.GetProcedureLineNo(match, code)
@@ -117,7 +144,7 @@ export class ALSyntaxUtil {
     }
 
     private static GetProcedureLineNo(procedureName: string, code: string): number {        
-        let codeLines = code.split("\r\n");
+        let codeLines = code.replace('\r','').split('\n');
         let pos: number = -1;
         codeLines.filter((x) => {
             if (x.includes(procedureName)) {
@@ -129,7 +156,8 @@ export class ALSyntaxUtil {
     }
 
     public static AnalyzeProcedureDefinition(code: string): RegExpMatchArray | null {
-        return this.AnalyzeDefinition(code.match(/(trigger|(?!local)procedure)\s+(?<ProcedureName>[A-Za-z0-9_]+)\b[^\(]*\((?<Params>.*)\)(?<ReturnType>((.*\:\s*)[A-Za-z0-9\s\""\.\[\]]+))?/));
+        //return this.AnalyzeDefinition(code.match(/(trigger|(?!local)procedure)\s+(?<ProcedureName>[A-Za-z0-9_]+)\b[^\(]*\((?<Params>.*)\)(?<ReturnType>((.*\:\s*)[A-Za-z0-9\s\""\.\[\]]+))?/));
+        return this.AnalyzeDefinition(code.match(/(?<!\/\/\/.*)(?<Type>(procedure|trigger))\s+(\")?(?<ProcedureName>.+)\b[^\(]*\((?<Params>.*)\)((\s*(return)?\:\s*)(?<ReturnType>[A-Za-z0-9\s\""\.\[\]]+))?/));
     }
 
     public static AnalyzeObjectDefinition(code: string): RegExpMatchArray | null {